+++ src/Sentry.Profiling/SampleProfilerSession.cs
@@ -16,14 +16,16 @@ internal class SampleProfilerSession : IDisposable
     private readonly IDiagnosticLogger? _logger;
     private readonly SentryStopwatch _stopwatch;
     private bool _stopped = false;
+    private Task _processing;
 
-    private SampleProfilerSession(SentryStopwatch stopwatch, EventPipeSession session, TraceLogEventSource eventSource, IDiagnosticLogger? logger)
+    private SampleProfilerSession(SentryStopwatch stopwatch, EventPipeSession session, TraceLogEventSource eventSource, Task processing, IDiagnosticLogger? logger)
     {
         _session = session;
         _logger = logger;
         EventSource = eventSource;
         _sampleEventParser = new SampleProfilerTraceEventParser(EventSource);
         _stopwatch = stopwatch;
+        _processing = processing;
     }
 
     // Exposed only for benchmarks.
@@ -88,7 +90,7 @@ internal class SampleProfilerSession : IDisposable
             var eventSource = TraceLog.CreateFromEventPipeSession(session, TraceLog.EventPipeRundownConfiguration.Enable(client));
 
             // Process() blocks until the session is stopped so we need to run it on a separate thread.
-            Task.Factory.StartNew(eventSource.Process, TaskCreationOptions.LongRunning)
+            var processing = Task.Factory.StartNew(eventSource.Process, TaskCreationOptions.LongRunning)
                 .ContinueWith(_ =>
                 {
                     if (_.Exception?.InnerException is { } e)
@@ -97,7 +99,7 @@ internal class SampleProfilerSession : IDisposable
                     }
                 }, TaskContinuationOptions.OnlyOnFaulted);
 
-            return new SampleProfilerSession(stopWatch, session, eventSource, logger);
+            return new SampleProfilerSession(stopWatch, session, eventSource, processing, logger);
         }
         catch (Exception ex)
         {
@@ -130,6 +132,7 @@ internal class SampleProfilerSession : IDisposable
             {
                 _stopped = true;
                 _session.Stop();
+                _processing.Wait();
                 _session.Dispose();
                 EventSource.Dispose();
             }